Split browser row window projection
